[VIBRA] Remove vibrator AOSP implementation from libhardware_legacy
To enable the vendor implementation of the vibrator HAL, it is needed
to use hardware modules, a default one for AOSP and the vendor one.
The implementation in libhardware_legacy is no more needed so remove
it from the file system.

Note: This patch is related to a change in platforms/hardware/libhardware

Change-Id: Ia7f7770bd52502d6f2c44be05f5a7d454db5994a
Author: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Author-tracking-BZ: 49760