[AUDIO] Support new audio devices enum values introduced in JB MR1 code drop
Today, each audio device is identified by a bit in a 32 bit field with
16 bits for output devices and 16 bits for input devices.  So the
number of output devices is now exhausted i.e. we can't add new output
devices.

So in order to add new devices, a new convention for audio device
identification is defined, allowing 31 different input and output
devices.  Bit 31 of the bit field is reserved and when set indicates
an input device.

Change-Id: If04c20c30889cc155f8ed3b3bcae7ba95fa8429c
Author: Neil di Folco <neilx.di.folco@intel.com>
Signed-off-by: Neil di Folco <neilx.di.folco@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Author-tracking-BZ: 71815