wifi: Allow P2P functionality to be disabled per interface.
By default, P2P is enabled globally for all virtual interfaces and this
makes wpa_supplicant include WSC and P2P IEs in Probe Request frames for
all scans even if this is for a non-P2P station connection to speed up
device discovery. If an interface is dedicated for non-P2P station mode
operations, it is now possible to disable addition of WSC and P2P IEs
into Probe Request frames with a per-interface p2p_disabled parameter.
This can be set in the configuration file (p2p_disabled=1).Hence set
p2p_disabled for wlan0 interface in the wpa_supplicant.conf.

Change-Id: If60502d75d602aa431aa308b1d5e2a03a976d83e